fix : Compare moved to pycm_compare #111
