+++ pycm/__init__.py
@@ -3,5 +3,6 @@ from .pycm_class_func import *
 from .pycm_overall_func import *
 from .pycm_output import *
 from .pycm_obj import *
+from .pycm_compare import *
 from .pycm_util import *
 from .pycm_interpret import *
